requirements.txt
django == 4.2.3
pillow
django-admin-interface
django-crispy-forms
django-rest-framework
